Lvl Up
Lvl up and stats of enemies connected to lvl ups

+++ Game_for_school/Assets/Scripts/EnemyData.cs
@@ -5,8 +5,8 @@ using UnityEngine;
 public class EnemyData : ScriptableObject
 {
     public string enemyName;
-    public int maxHealth;
-    public int damage;
+    public float maxHealth;
+    public float damage;
     public string attackType;
     public Sprite portrait;
     public GameObject enemyPrefab;
